Break Record to own file
